% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{batch_chat}
\alias{batch_chat_structured}
\alias{batch_chat_text}
\alias{batch_chat_completed}
\alias{parallel_chat}
\alias{parallel_chat_text}
\alias{parallel_chat_structured}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ellmer}{\code{\link[ellmer]{batch_chat}}, \code{\link[ellmer:batch_chat]{batch_chat_completed}}, \code{\link[ellmer:batch_chat]{batch_chat_structured}}, \code{\link[ellmer:batch_chat]{batch_chat_text}}, \code{\link[ellmer]{parallel_chat}}, \code{\link[ellmer:parallel_chat]{parallel_chat_structured}}, \code{\link[ellmer:parallel_chat]{parallel_chat_text}}}
}}

