% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-groq.R
\name{set_default}
\alias{set_default}
\title{Set default value with informative message}
\usage{
set_default(value, default, arg = rlang::caller_arg(value))
}
\description{
Set default value with informative message
}
\keyword{internal}
